Moved distillation model to distillation.py for code consistency.
